// packages/stable/core/src/types/interfaces/i-document-data.ts
export interface IDocumentBody {
    dataStream: string;

    textRuns?: ITextRun[]; // textRun style，interaction

    paragraphs?: IParagraph[]; // paragraph

    sectionBreaks?: ISectionBreak[]; // SectionBreak https://support.microsoft.com/en-us/office/insert-a-section-break-eef20fd8-e38c-4ba6-a027-e503bdf8375c

    customBlocks?: ICustomBlock[]; // customBlock user-defined block through plug-in

    tables?: ICustomTable[]; // Table

    // tableOfContents?: { [index: number]: ITableOfContent }; // tableOfContents
    // links?: { [index: number]: IHyperlink }; // links

    customRanges?: ICustomRange[]; // plugin register，implement special logic for streams， hyperlink, field，structured document tags， bookmark，comment

    customDecorations?: ICustomDecoration[];

    /**
     * for copy/paste, data of custom-range and other module
     * it won't save to disk
     */
    payloads?: Record<string, string>;
}
// ...
export interface ITextRun {
    // ct?: string; // content
    // len: number;
    st: number;
    ed: number;
    sId?: string; // styleID
    ts?: ITextStyle; // textStyle
    // tab?: BooleanNumber; // if tab，default is false
}

/**
 * Block element, link like, disabled to self nested
 */
export interface ICustomRange<T extends Record<string, any> = Record<string, any>> {
    startIndex: number;
    endIndex: number;
    rangeId: string;
    rangeType: CustomRangeType | number;
    /**
     * display as a whole-entity
     */
    wholeEntity?: boolean;
    /**
     * properties of custom range,
     * for example, hyperlink: `{ url: string }`
     */
    properties?: T;
}

export type IHyperLinkCustomRange = ICustomRange<{ url: string }>;

export type IMentionCustomRange = ICustomRange<IMention>;

export enum CustomRangeType {
    HYPERLINK,
    FIELD, // 17.16 Fields and Hyperlinks
    SDT, // 17.5.2 Structured Document Tags
    BOOKMARK,
    COMMENT,
    CUSTOM,
    MENTION,
    UNI_FORMULA,

    DELTED = 9999,
}

/**
 * Custom Block
 */
export interface ICustomBlock {
    startIndex: number;
    blockType?: BlockType;
    // A unique ID associated with a custom block.
    blockId: string;
}

export enum CustomDecorationType {
    COMMENT,
    DELETED = 9999,
}

export interface ICustomDecoration {
    startIndex: number;
    endIndex: number;
    id: string;
    type: CustomDecorationType;
}

/**
 * Type of block
 */
export enum BlockType {
    DRAWING,
    CUSTOM,
}

export interface IHeaderAndFooterBase {
    defaultHeaderId?: string; // defaultHeaderId or the odd page header id
    defaultFooterId?: string; // defaultFooterId or the odd page footer id
    evenPageHeaderId?: string; // evenPageHeaderId
    evenPageFooterId?: string; // evenPageFooterId
    firstPageHeaderId?: string; // firstPageHeaderId
    firstPageFooterId?: string; // firstPageFooterId
    useFirstPageHeaderFooter?: BooleanNumber; // useFirstPageHeaderFooter
    evenAndOddHeaders?: BooleanNumber; // useEvenPageHeaderFooter,
}

export enum DocumentFlavor {
    UNSPECIFIED,
    TRADITIONAL,
    MODERN,
}

/**
 * Basics properties of doc style
 */
export interface IDocStyleBase extends IMargin {
    pageNumberStart?: number; // pageNumberStart
    pageSize?: ISize; // pageSize

    pageOrient?: PageOrientType;

    documentFlavor?: DocumentFlavor; // DocumentFlavor: TRADITIONAL, MODERN

    marginHeader?: number; // marginHeader
    marginFooter?: number; // marginFooter

    renderConfig?: IDocumentRenderConfig;
}
// ...
export interface ISectionBreakBase {
    // docGrid (Document Grid), open xml $17.6.5
    charSpace?: number; // charSpace
    linePitch?: number; // linePitch
    gridType?: GridType; // gridType

    columnProperties?: ISectionColumnProperties[]; // columnProperties 17.6.4 cols (Column Definitions)
    columnSeparatorType?: ColumnSeparatorType; // ColumnSeparatorType
    contentDirection?: TextDirection; // contentDirection
    sectionType?: SectionType; // sectionType 17.6.22 type (Section Type)
    // deprecated: The attribute does not exist in Word and should be deprecated.
    sectionTypeNext?: SectionType; // sectionType
    textDirection?: TextDirectionType; // textDirection
}

/**
 * Properties of section break
 */
export interface ISectionBreak extends IDocStyleBase, ISectionBreakBase, IHeaderAndFooterBase {
    startIndex: number;
}
// ...
export interface IParagraph {
    // elements: IElement[]; // elements
    startIndex: number;
    paragraphStyle?: IParagraphStyle; // paragraphStyle
    bullet?: IBullet; // bullet
    // dIds?: string[]; // drawingIds drawingId
}
// ...
export interface IBullet {
    listType: string; // listType orderList or bulletList etc.
    listId: string; // listId
    nestingLevel: number; // nestingLevel
    textStyle?: ITextStyle; // textStyle
}
// ...
export interface ITextStyle extends IStyleBase {
    // bo?: BaselineOffset; // BaselineOffset, sup, sub
    sc?: number; // spacing
    pos?: number; // position
    sa?: number; // scale
}

export interface IIndentStart {
    indentFirstLine?: INumberUnit; // indentFirstLine，17.3.1.12 ind (Paragraph Indentation)
    hanging?: INumberUnit; // hanging，offset of first word except first line
    indentStart?: INumberUnit; // indentStart
    tabStops?: ITabStop[]; // tabStops
    indentEnd?: INumberUnit; // indentEnd
}

/**
 * Properties of paragraph style
 */
export interface IParagraphStyle extends IParagraphProperties {
    // Not achieved, aligned with Excel's standards.
    textStyle?: ITextStyle; // paragraph textStyle
}
// ...
export interface ICustomTable {
    startIndex: number;
    endIndex: number;
    // A unique ID associated with a table.
    tableId: string;
}